cargo: change name back to "sysfs_gpio"
Renaming the package on crates.io is not currently supported
in any way that I can determine.  Users can download the
package as "sysfs-gpio" but it will still show up as sysfs_gpio
on the site.  It is what it is...

Move version back down to 0.3.1 as there are no breaking changes
in this release as there is no major name change.