Add `from_path` constructor to support symlinks
A common pattern in production embedded systems is to export
pins at startup with symlinks to a given location.  Previously,
there was not a great way to work with this mode here as we
expected the caller to know the pin number (and not just a
symlink containing a name).

Signed-off-by: Paul Osborne <osbpau@gmail.com>