Loosen strictness of Pin::from_path regex
Previously the Pin::from_path function used a regex that looked
explicitly for a file in /sys/class/gpio/gpio(\d+). The problem with this is
that the gpios themselves may be kept in a symlinked directory -
especially when the GPIOs are resolved to a System on Chip (SoC).

This commit adds a lazy match to the middle of the regex check - the
path must still start with /sys and end with /gpio/gpio(\d+), but the
intermediate paths may be ignored.

Signed-off-by: Nick Stevens <nick@bitcurry.com>